refactor unit tests, remove import cycle
